@@ -8,6 +8,11 @@ namespace Xamarin.CommunityToolkit.Converters
 	public class MathExpressionConverter : ValueConverterExtension, IValueConverter
 	{
 		/// <summary>
+		/// The expression to calculate.
+		/// </summary>
+		public string? Expression { get; set; }
+
+		/// <summary>
 		/// Calculate the incoming expression string with one variable.
 		/// </summary>
 		/// <param name="value">The variable X for an expression</param>
@@ -17,7 +22,7 @@ namespace Xamarin.CommunityToolkit.Converters
 		/// <returns>A <see cref="double"/> The result of calculating an expression.</returns>
 		public object? Convert(object? value, Type? targetType, object? parameter, CultureInfo culture)
 		{
-			if (parameter is not string expression)
+			if ((parameter ?? Expression) is not string expression)
 				throw new ArgumentException("The parameter should be of type String.");
 
 			if (value == null || !double.TryParse(value.ToString(), out var xValue))
