@@ -1,11 +1,11 @@
-﻿using System.Collections.ObjectModel;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Linq;
 using CoreGraphics;
 using Foundation;
 using UIKit;
 using Xamarin.CommunityToolkit.UI.Views;
-using Xamarin.CommunityToolkit.UI.Views.iOS;
 using Xamarin.Forms;
 using Xamarin.Forms.Platform.iOS;
 
@@ -18,11 +18,12 @@ namespace Xamarin.CommunityToolkit.UI.Views
 	/// </summary>
 	public class DrawingViewRenderer : ViewRenderer<DrawingView, UIView>
 	{
+		readonly UIBezierPath currentPath;
 		bool disposed;
-		UIBezierPath currentPath;
 		UIColor? lineColor;
 		CGPoint previousPoint;
 		Line? currentLine;
+		readonly List<ScrollViewRenderer> scrollViewParentRenderers = new();
 
 		public DrawingViewRenderer() => currentPath = new UIBezierPath();
 
@@ -49,6 +50,7 @@ namespace Xamarin.CommunityToolkit.UI.Views
 
 		public override void TouchesBegan(NSSet touches, UIEvent? evt)
 		{
+			DetectScrollViews();
 			SetParentTouches(false);
 
 			Element.Lines.CollectionChanged -= OnLinesCollectionChanged;
@@ -61,11 +63,11 @@ namespace Xamarin.CommunityToolkit.UI.Views
 			var touch = (UITouch)touches.AnyObject;
 			previousPoint = touch.PreviousLocationInView(this);
 			currentPath.MoveTo(previousPoint);
-			currentLine = new Line()
+			currentLine = new Line
 			{
-				Points = new ObservableCollection<Point>()
+				Points = new ObservableCollection<Point>
 				{
-					new Point(previousPoint.X, previousPoint.Y)
+					new (previousPoint.X, previousPoint.Y)
 				}
 			};
 
@@ -79,6 +81,7 @@ namespace Xamarin.CommunityToolkit.UI.Views
 			var touch = (UITouch)touches.AnyObject;
 			var currentPoint = touch.LocationInView(this);
 			AddPointToPath(currentPoint);
+			SetNeedsDisplay();
 			currentLine?.Points.Add(currentPoint.ToPoint());
 		}
 
@@ -86,7 +89,6 @@ namespace Xamarin.CommunityToolkit.UI.Views
 		{
 			if (currentLine != null)
 			{
-				UpdatePath(currentLine);
 				Element.Lines.Add(currentLine);
 				Element.OnDrawingLineCompleted(currentLine);
 			}
@@ -94,34 +96,34 @@ namespace Xamarin.CommunityToolkit.UI.Views
 			if (Element.ClearOnFinish)
 				Element.Lines.Clear();
 
+			currentLine = null;
 			SetParentTouches(true);
 		}
 
 		public override void TouchesCancelled(NSSet touches, UIEvent? evt)
 		{
-			InvokeOnMainThread(SetNeedsDisplay);
+			currentLine = null;
+			SetNeedsDisplay();
 			SetParentTouches(true);
 		}
 
 		public override void Draw(CGRect rect)
 		{
-			lineColor!.SetStroke();
+			lineColor?.SetStroke();
 			currentPath.Stroke();
 		}
 
-		void AddPointToPath(CGPoint currentPoint)
-		{
-			currentPath.AddLineTo(currentPoint);
-			SetNeedsDisplay();
-		}
+		void AddPointToPath(CGPoint currentPoint) => currentPath.AddLineTo(currentPoint);
 
 		void LoadPoints()
 		{
 			currentPath.RemoveAllPoints();
 			foreach (var line in Element.Lines)
 			{
-				UpdatePath(line);
-				var stylusPoints = line.Points.Select(point => new CGPoint(point.X, point.Y)).ToList();
+				var newPointsPath = line.EnableSmoothedPath
+					? line.Points.SmoothedPathWithGranularity(line.Granularity)
+					: line.Points;
+				var stylusPoints = newPointsPath.Select(point => new CGPoint(point.X, point.Y)).ToList();
 				if (stylusPoints.Count > 0)
 				{
 					previousPoint = stylusPoints[0];
@@ -133,82 +135,6 @@ namespace Xamarin.CommunityToolkit.UI.Views
 
 			SetNeedsDisplay();
 		}
-
-		void UpdatePath(Line line)
-		{
-			Element.Lines.CollectionChanged -= OnLinesCollectionChanged;
-			var smoothedPoints = line.EnableSmoothedPath
-				? SmoothedPathWithGranularity(line.Points, line.Granularity)
-				: new ObservableCollection<Point>(line.Points);
-
-			line.Points.Clear();
-
-			foreach (var point in smoothedPoints)
-				line.Points.Add(point);
-
-			Element.Lines.CollectionChanged += OnLinesCollectionChanged;
-		}
-
-		ObservableCollection<Point> SmoothedPathWithGranularity(ObservableCollection<Point> currentPoints,
-			int granularity)
-		{
-			// not enough points to smooth effectively, so return the original path and points.
-			if (currentPoints.Count < granularity + 2)
-				return new ObservableCollection<Point>(currentPoints);
-
-			var smoothedPoints = new ObservableCollection<Point>();
-
-			// duplicate the first and last points as control points.
-			currentPoints.Insert(0, currentPoints[0]);
-			currentPoints.Add(currentPoints[^1]);
-
-			// add the first point
-			smoothedPoints.Add(currentPoints[0]);
-
-			var currentPointsCount = currentPoints.Count;
-			for (var index = 1; index < currentPointsCount - 2; index++)
-			{
-				var p0 = currentPoints[index - 1];
-				var p1 = currentPoints[index];
-				var p2 = currentPoints[index + 1];
-				var p3 = currentPoints[index + 2];
-
-				// add n points starting at p1 + dx/dy up until p2 using Catmull-Rom splines
-				for (var i = 1; i < granularity; i++)
-				{
-					var t = i * (1f / granularity);
-					var tt = t * t;
-					var ttt = tt * t;
-
-					// intermediate point
-					var mid = GetIntermediatePoint(p0, p1, p2, p3, t, tt, ttt);
-					smoothedPoints.Add(mid);
-				}
-
-				// add p2
-				smoothedPoints.Add(p2);
-			}
-
-			// add the last point
-			var last = currentPoints[^1];
-			smoothedPoints.Add(last);
-			return smoothedPoints;
-		}
-
-		Point GetIntermediatePoint(Point p0, Point p1, Point p2, Point p3, in float t, in float tt, in float ttt) =>
-			new Point
-			{
-				X = 0.5f *
-					((2f * p1.X) +
-					 ((p2.X - p0.X) * t) +
-					 (((2f * p0.X) - (5f * p1.X) + (4f * p2.X) - p3.X) * tt) +
-					 (((3f * p1.X) - p0.X - (3f * p2.X) + p3.X) * ttt)),
-				Y = 0.5f *
-					((2 * p1.Y) +
-					 ((p2.Y - p0.Y) * t) +
-					 (((2 * p0.Y) - (5 * p1.Y) + (4 * p2.Y) - p3.Y) * tt) +
-					 (((3 * p1.Y) - p0.Y - (3 * p2.Y) + p3.Y) * ttt))
-			};
 
 		protected override void Dispose(bool disposing)
 		{
@@ -227,16 +153,27 @@ namespace Xamarin.CommunityToolkit.UI.Views
 			base.Dispose(disposing);
 		}
 
-		void SetParentTouches(bool enabled)
+		void DetectScrollViews()
 		{
+			if (scrollViewParentRenderers.Any())
+				return;
+
 			var parent = Superview;
 
 			while (parent != null)
 			{
 				if (parent.GetType() == typeof(ScrollViewRenderer))
-					((ScrollViewRenderer)parent).ScrollEnabled = enabled;
+					scrollViewParentRenderers.Add((ScrollViewRenderer)parent);
 
 				parent = parent.Superview;
+			}
+		}
+
+		void SetParentTouches(bool enabled)
+		{
+			foreach (var scrollViewParentRenderer in scrollViewParentRenderers)
+			{
+				scrollViewParentRenderer.ScrollEnabled = enabled;
 			}
 		}
 	}
